skills-dev/cost/requirements.txt
# 工程造价 Skill 依赖包

# 核心依赖
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0  # PostgreSQL 驱动

# CAD 处理
ezdxf>=1.1.0  # DXF/DWG 文件读取
matplotlib>=3.7.0  # CAD 可视化

# 视觉分析
openai>=1.0.0  # OpenAI SDK (兼容 Kimi 等)
Pillow>=10.0.0  # 图像处理

# Excel 导出
openpyxl>=3.1.0  # Excel 文件操作

# 向量搜索（可选）
pgvector>=0.2.0  # PostgreSQL 向量扩展

# 网络爬虫（可选）
requests>=2.31.0
beautifulsoup4>=4.12.0

# 工具库
python-dotenv>=1.0.0  # 环境变量管理
